corrected bug on pagination and moved BrapiObservationDTO class
